Fixes for Summit
Fix some typos and grammatical errors.

Make some fixes to example output to bring it a little closer to what the
current versions of tools actually output (some discrepancies remain).

Add some instructions on connecting to the virtual machines that we are
using for the Summit lab.

Be sure to tell the reader both usernames and passwords for the accounts
that they will be using.

Restrict the set of cartridges that we install on the node to reduce
installation time.

Add verification of the broker's REST API, using curl to issue an
authenticated request.  Doing so has the side effect of ensuring that the
OpenShift broker will add the 'demo' user to MongoDB.  Without this step,
the oo-admin-ctl-user commands in the following chapter would fail.

Tell the user to use su instead of sudo because the lab machines do not
allow sudo access, but we can give away the root password.

Remove one last instance of smartquotes.